Implementing a "Udp-handshake" for the audio-bridge.
Implementing the "udp-handshake". As well as modifying some typedefs. Also modified the StopWatch class.

To ensure that it can connect to itself, even while the client and server is running on the same machine, without feedbacking into itself.

Modifying the StopWatch class to be lockless.

Removing a superfluous "index" from the DatagramSocketClient's processReadData function.

Modifying the reconnection logic of the AudioBridge.

Updating the timing mechanism of the VoiceConnection class.